+++ src/hardware/booster_channels.rs
@@ -10,8 +10,7 @@ use stm32f4xx_hal as hal;
 use tca9548::{self, Tca9548};
 
 use super::rf_channel::{ChannelPins as RfChannelPins, RfChannel, RfChannelMachine};
-use super::{Channel, I2cBusManager, I2cProxy, SystemTimer};
-use embedded_hal::blocking::delay::DelayUs;
+use super::{delay::AsmDelay, Channel, I2cBusManager, I2cProxy, SystemTimer};
 
 /// Represents a control structure for interfacing to booster RF channels.
 pub struct BoosterChannels {
@@ -56,7 +55,7 @@ impl BoosterChannels {
         manager: &'static I2cBusManager,
         pins: [RfChannelPins; 8],
         clock: SystemTimer,
-        delay: &mut impl DelayUs<u16>,
+        delay: AsmDelay,
     ) -> Self {
         let mut channels: [Option<RfChannelMachine>; 8] =
             [None, None, None, None, None, None, None, None];
@@ -66,7 +65,7 @@ impl BoosterChannels {
             mux.select_bus(Some(idx.into()))
                 .expect("Failed to select channel");
 
-            if let Some(channel) = RfChannel::new(manager, pins, clock, delay) {
+            if let Some(channel) = RfChannel::new(manager, pins, clock, delay.clone()) {
                 let mut machine = RfChannelMachine::new(channel);
                 machine.handle_startup();
                 channels[idx as usize].replace(machine);