+++ src/hardware/rf_channel.rs
@@ -12,14 +12,13 @@ use mcp3221::Mcp3221;
 use microchip_24aa02e48::Microchip24AA02E48;
 use minimq::embedded_time::{duration::Extensions, Clock, Instant};
 
-use super::{platform, I2cBusManager, I2cProxy, SystemTimer};
+use super::{delay::AsmDelay, platform, I2cBusManager, I2cProxy, SystemTimer};
 use crate::{
     settings::{
         channel_settings::ChannelSettings, channel_settings::ChannelState, BoosterChannelSettings,
     },
     Error,
 };
-use embedded_hal::blocking::delay::DelayUs;
 use stm32f4xx_hal::{
     self as hal,
     adc::config::SampleTime,
@@ -157,7 +156,7 @@ impl Devices {
     /// properly enumerate, the option will be empty. The returend tuple will be (devices, eeprom).
     fn new(
         manager: &'static I2cBusManager,
-        delay: &mut impl DelayUs<u16>,
+        delay: &mut AsmDelay,
     ) -> Option<(Self, Microchip24AA02E48<I2cProxy>)> {
         // The ADS7924 and DAC7571 are present on the booster mainboard, so instantiation
         // and communication should never fail.
@@ -281,6 +280,7 @@ pub struct RfChannel {
     pins: ChannelPins,
     settings: BoosterChannelSettings,
     clock: SystemTimer,
+    delay: AsmDelay,
 }
 
 impl RfChannel {
@@ -300,15 +300,16 @@ impl RfChannel {
         manager: &'static I2cBusManager,
         pins: ChannelPins,
         clock: SystemTimer,
-        delay: &mut impl DelayUs<u16>,
+        mut delay: AsmDelay,
     ) -> Option<Self> {
         // Attempt to instantiate the I2C devices on the channel.
-        Devices::new(manager, delay).map(|(devices, eeprom)| {
+        Devices::new(manager, &mut delay).map(|(devices, eeprom)| {
             let mut channel = Self {
                 devices,
                 pins,
                 settings: BoosterChannelSettings::new(eeprom),
                 clock,
+                delay,
             };
 
             channel.apply_output_interlock_threshold().unwrap();
@@ -671,8 +672,18 @@ impl sm::StateMachineContext for RfChannel {
             .set_voltage(3.2)
             .expect("Failed to disable RF bias voltage");
 
-        // Ensure that the RF output is disabled during the power-up process.
-        self.disable_rf_switch();
+        // Next, handle resetting interlocks for v1.6 hardware. The interlocks are reset by a
+        // falling edge on ON/OFF. Because the bias dac is currently in pinch-off (and the RF
+        // channel is unpowered), toggling ON/OFF introduces no output transients on the RF
+        // connectors.
+
+        self.pins.signal_on.set_high().unwrap();
+
+        // Note: The delay here are purely to accomodate potential capacitance on the ON/OFF
+        // rails.
+        self.delay.delay_ms(1u32);
+
+        self.pins.signal_on.set_low().unwrap();
 
         // Start the LM3880 power supply sequencer.
         self.pins.enable_power.set_high().unwrap();
@@ -737,6 +748,8 @@ impl sm::StateMachineContext for RfChannel {
         assert!(settings.output_interlock_threshold > settings.output_power_transform.map(0.100));
 
         self.apply_bias().unwrap();
+
+        // Ensure that the final state is an enabled ON signal.
         self.pins.signal_on.set_high().unwrap();
     }